requirements.txt
gradio==5.14.0
kagglehub==0.3.7
langchain-chroma==0.2.1
langchain-community==0.3.16
langchain-huggingface==0.1.2
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langchain-core==0.3.33

matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
seaborn==0.13.2
tqdm==4.67.1
transformers==4.48.2
